+++ src/Chats/Messages.php
@@ -12,6 +12,28 @@ use Swe\SpaceSDK\Exception\MissingArgumentException;
 class Messages extends AbstractApi
 {
     /**
+     * Delete a message from a channel.
+     *
+     * Permissions that may be checked: Channel.PostMessage, Channel.PostMessagesInThreads
+     *
+     * @param array $data
+     * @return void
+     * @throws GuzzleException
+     * @throws MissingArgumentException
+     */
+    public function deleteMessage(array $data): void
+    {
+        $uri = 'chats/messages/delete-message';
+        $required = [
+            'channel' => 'string',
+            'id' => 'string',
+        ];
+        $this->throwIfInvalid($required, $data);
+
+        $this->client->post($this->buildUrl($uri), $data);
+    }
+
+    /**
      * This API method is intended to be used only by applications.
      *
      * Permissions that may be checked: Channel.ImportMessages