+++ changelog.md
@@ -7,6 +7,7 @@
 - Added method `getChannelAdministrator` in `Administrator`
 - Added class `Chats/Channels/Description`
 - Added method `changeChannelDescription` in `Description`
+- Added method `deleteMessage` in `Messages`
 - **[BREAKING]** Renamed method and class `GraphExecution` to `GraphExecutions`
 - **[BREAKING]** Renamed method and class `Job` to `Jobs`
 - **[BREAKING]** Renamed method and class `Repository` to `Repositories`